// frontend/src/api/generated/shop/schemas/productDelivery.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * ShopDjango API
 * API for ShopDjango project
 * OpenAPI spec version: 1.0.0
 */

export interface ProductDelivery {
  readonly id: number;
  /** Supplier who delivered the products */
  supplier: number;
  /** Product that was delivered */
  product: number;
  /**
   * Quantity of products delivered
   * @minimum 1
   * @maximum 2147483647
   */
  quantity: number;
  /** When the delivery was received */
  delivery_date: string;
  /**
   * Cost per unit for this delivery
   * @pattern ^-?\d{0,8}(?:\.\d{0,2})?$
   */
  cost_per_unit: string;
  /** Timestamp when the record was created */
  readonly created_at: string;
  /** Timestamp when the record was last updated */
  readonly updated_at: string;
}